refactor: shared error middleware for generic errors
- added shared error middleware for generic errors related to auth, request validation, jwt, not found, and mongoose
- refactored service-specific middlewares to extend the shared error middleware
