+++ shared/shared-common/src/middlewares/index.ts
@@ -1,1 +1,2 @@
 export * from './validator.middleware';
+export * from './error.middleware';
